refactor(media-library): extract marquee selection into a hook
Pulls the ~115-line marquee preview/selection cluster (DOM-level preview
highlighting, marquee item rects, useMarqueeSelection wiring, store commit) out
of the 1882-line MediaLibrary component into use-media-library-marquee.ts.
Verbatim; stable scrollContainerRef added to dep arrays.

media-library.tsx: 1882 -> 1771 lines.
